fix(rpc): explicitly add locals to RPC context
Prior to this commit, globals from the `main.main()` function were added
to the `context` dictionary that was passed to the `rpc.rpc()`
coroutine. However, this resulted in a recursive `rpc` reference within
the `rpc.rpc()` function. Moreover, the locals from the `main.main()`
function were explicitly added _directly to the `globals()` dictionary.

In this commit:

 1. Remove the `rpc` module from the `context` dictionary to prevent the
    recursive module reference in `rpc.rpc()`.
 2. Explicitly add the local variables from `main.main()` to the
    `context` dictionary.

It is worth noting that **(2)** is necessary (as opposed to just
updating the `context` dictionary with the contents of `locals()`)
since (see [here][1]):

> MicroPython optimizes local variable handling and does not record or
> provide any introspection info for them, e.g. **locals() doesn't have
> entries for locals.**

[1]: https://github.com/micropython/micropython/wiki/Differences#differences-by-design