fix(config): only try to read if file exists
Prior to this commit, an attempt was made to read `config.json` on boot
up. However, this would fail in the case where the `config.json` file
did not exist (e.g., after initial bootstrap of ESP32 device).

In this commit, only attempt to read the configuration from
`config.json` if the file exists.