fix(wifimgr): do not automatically start webserver
Prior to this commit, the `wifimgr.get_connection()` function (executed on
boot) would automatically configure a wifi access point and launch a webserver
to configure wifi credentials for one of the detected networks. However,
this made it impossible to continue booting without configuring a valid wifi
connection.

In this commit, do not start the wifi configuration webserver if a connection
cannot be established. Instead, leave it up to the calling code to execute
`wifimgr.start()` as needed to start the wifi configuration webserver
explicitly; e.g., by holding a push button during boot, by sending a RPC
command, etc.

Fixes #1.