feat(notebooks/pump): vid/pid for port detection
Prior to this commit, the first available serial port was used assumed
to be connected to UART2 on the ESP32. However, this may not always be
the case.

In this commit, modify the pump UI notebook to select the appropriate
serial port based on the USB vendor ID and product ID. This should
connect to the correct port regardless of the assigned port number.