fix(notebook/huzzah): disable flush
Prior to this commit, the pump UI notebook explicitly flushed the
incoming serial buffer after connecting to the ESP32. This seemed to be
necessary when running on the Lolin32 ESP32 board since resetting the
board would send "garbage" over UART2. However, this does not seem to be
an issue when running on the Huzzah32 ESP32 board.

In this commit, do not flush the serial interface.