feat(ui): add task-based wash, elution, lysate UI
Prior to this commit, all pump and valve control was explicitly manual.
However, this requires manual mapping between desired flow path and the
corresponding pump addresses and valve address and states.

In this commit, introduce a separate set of task-based controls. As of
this commit, allow the following tasks to be controlled using only a
pump button, a slider to indicate the number of pulses, and a slider
indicating the pulse period:

- Pump wash buffer to beads
- Pump lysate to beads
- Pump elution to beads